+++ tests/stacks-collect-process.test.ts
@@ -1,364 +1,6 @@
-import { describe, test, expect, mock } from 'bun:test';
+import { describe, test, expect } from 'bun:test';
 import { spawn } from 'child_process';
-
-/**
- * stacks.ts imports `./db` (src/lib/server/db.ts), whose companion module
- * `./db/drizzle` runs `await seedDatabase()` at import time. That opens a
- * real better-sqlite3 database -- and the native binding does not load under
- * Bun (ERR_DLOPEN_FAILED, see https://github.com/oven-sh/bun/issues/4290,
- * confirmed against this checkout: `node -e "require('better-sqlite3')"`
- * succeeds, `bun -e` does not). Importing stacks.ts for real therefore
- * crashes before this file's tests ever run -- and, worse, the crash is
- * cached in Bun's process-global module registry and leaks into OTHER,
- * unrelated test files that run in the same `bun test tests/` invocation
- * (reproduced: it broke the pre-existing tests/semver-generic-v2.test.ts).
- *
- * collectProcess itself has no DB dependency, so `./db` is mocked here --
- * mirroring the established pattern in tests/backups/route-guards.test.ts
- * ("The module transitively imports the DB/engine via ./index, so we
- * mock.module() ... BEFORE importing it"). Unlike that file's single-function
- * mock, stacks.ts's import graph (secretproviders/scheduler/notifications/git,
- * 30+ files repo-wide import ./db) pulls in a much larger surface, so every
- * export db.ts currently declares is stubbed -- Bun's static ESM binding
- * check rejects a mock object missing ANY name a transitively-loaded module
- * imports, one at a time (confirmed: a partial/Proxy-based stub fails the
- * same way). Regenerate this list if db.ts's export surface changes:
- *
- *   python3 -c "
- *   import re
- *   names, seen = [], set()
- *   for line in open('src/lib/server/db.ts'):
- *       m = re.match(r'^export (async function|function|const) ([A-Za-z0-9_]+)', line)
- *       if m and m.group(2) not in seen:
- *           seen.add(m.group(2)); names.append(m.group(2))
- *   print(len(names))"
- */
-mock.module('../src/lib/server/db/drizzle', () => ({
-	isPostgres: false,
-	isSqlite: true,
-	checkSchemaHealth: async (..._args: any[]): Promise<any> => undefined,
-	db: {} as any,
-	rawClient: {} as any,
-	environments: {} as any,
-	hawserTokens: {} as any,
-	registries: {} as any,
-	settings: {} as any,
-	stackEvents: {} as any,
-	hostMetrics: {} as any,
-	configSets: {} as any,
-	autoUpdateSettings: {} as any,
-	notificationSettings: {} as any,
-	environmentNotifications: {} as any,
-	authSettings: {} as any,
-	users: {} as any,
-	sessions: {} as any,
-	ldapConfig: {} as any,
-	oidcConfig: {} as any,
-	roles: {} as any,
-	userRoles: {} as any,
-	gitCredentials: {} as any,
-	gitRepositories: {} as any,
-	gitStacks: {} as any,
-	secretProviders: {} as any,
-	stackSources: {} as any,
-	containerIconOverrides: {} as any,
-	vulnerabilityScans: {} as any,
-	auditLogs: {} as any,
-	containerEvents: {} as any,
-	userPreferences: {} as any,
-	scheduleExecutions: {} as any,
-	stackEnvironmentVariables: {} as any,
-	pendingContainerUpdates: {} as any,
-	apiTokens: {} as any,
-	backupDestinations: {} as any,
-	backupConfigs: {} as any,
-	eq: (..._args: any[]): any => ({}),
-	and: (..._args: any[]): any => ({}),
-	or: (..._args: any[]): any => ({}),
-	desc: (..._args: any[]): any => ({}),
-	asc: (..._args: any[]): any => ({}),
-	like: (..._args: any[]): any => ({}),
-	sql: (..._args: any[]): any => ({}),
-	inArray: (..._args: any[]): any => ({}),
-	isNull: (..._args: any[]): any => ({}),
-	isNotNull: (..._args: any[]): any => ({}),
-	getDatabaseSchemaVersion: async (..._args: any[]): Promise<any> => undefined,
-	getPostgresConnectionInfo: async (..._args: any[]): Promise<any> => undefined,
-	getDatabaseStats: async (..._args: any[]): Promise<any> => undefined,
-}));
-
-mock.module('../src/lib/server/db', () => ({
-	initDatabase: async (..._args: any[]): Promise<any> => undefined,
-	getEnvironments: async (..._args: any[]): Promise<any> => undefined,
-	hasEnvironments: async (..._args: any[]): Promise<any> => undefined,
-	getEnvironment: async (..._args: any[]): Promise<any> => undefined,
-	getEnvironmentByName: async (..._args: any[]): Promise<any> => undefined,
-	createEnvironment: async (..._args: any[]): Promise<any> => undefined,
-	updateEnvironment: async (..._args: any[]): Promise<any> => undefined,
-	deleteEnvironment: async (..._args: any[]): Promise<any> => undefined,
-	getRegistries: async (..._args: any[]): Promise<any> => undefined,
-	getRegistry: async (..._args: any[]): Promise<any> => undefined,
-	getDefaultRegistry: async (..._args: any[]): Promise<any> => undefined,
-	createRegistry: async (..._args: any[]): Promise<any> => undefined,
-	updateRegistry: async (..._args: any[]): Promise<any> => undefined,
-	deleteRegistry: async (..._args: any[]): Promise<any> => undefined,
-	setDefaultRegistry: async (..._args: any[]): Promise<any> => undefined,
-	getSecretProviders: async (..._args: any[]): Promise<any> => undefined,
-	getSecretProviderById: async (..._args: any[]): Promise<any> => undefined,
-	createSecretProvider: async (..._args: any[]): Promise<any> => undefined,
-	updateSecretProvider: async (..._args: any[]): Promise<any> => undefined,
-	deleteSecretProvider: async (..._args: any[]): Promise<any> => undefined,
-	logStackEvent: async (..._args: any[]): Promise<any> => undefined,
-	getStackEvents: async (..._args: any[]): Promise<any> => undefined,
-	getSetting: async (..._args: any[]): Promise<any> => undefined,
-	setSetting: async (..._args: any[]): Promise<any> => undefined,
-	deleteSetting: async (..._args: any[]): Promise<any> => undefined,
-	listSettingsByPrefix: async (..._args: any[]): Promise<any> => undefined,
-	getEnvSetting: async (..._args: any[]): Promise<any> => undefined,
-	setEnvSetting: async (..._args: any[]): Promise<any> => undefined,
-	getUserSetting: async (..._args: any[]): Promise<any> => undefined,
-	setUserSetting: async (..._args: any[]): Promise<any> => undefined,
-	getUserThemePreferences: async (..._args: any[]): Promise<any> => undefined,
-	setUserThemePreferences: async (..._args: any[]): Promise<any> => undefined,
-	getGridPreferences: async (..._args: any[]): Promise<any> => undefined,
-	setGridPreferences: async (..._args: any[]): Promise<any> => undefined,
-	deleteGridPreferences: async (..._args: any[]): Promise<any> => undefined,
-	resetAllGridPreferences: async (..._args: any[]): Promise<any> => undefined,
-	getSidebarPreferences: async (..._args: any[]): Promise<any> => undefined,
-	setSidebarPreferences: async (..._args: any[]): Promise<any> => undefined,
-	deleteSidebarPreferences: async (..._args: any[]): Promise<any> => undefined,
-	getEnvironmentPublicIps: async (..._args: any[]): Promise<any> => undefined,
-	setEnvironmentPublicIp: async (..._args: any[]): Promise<any> => undefined,
-	deleteEnvironmentPublicIp: async (..._args: any[]): Promise<any> => undefined,
-	getConfigSets: async (..._args: any[]): Promise<any> => undefined,
-	getConfigSet: async (..._args: any[]): Promise<any> => undefined,
-	createConfigSet: async (..._args: any[]): Promise<any> => undefined,
-	updateConfigSet: async (..._args: any[]): Promise<any> => undefined,
-	deleteConfigSet: async (..._args: any[]): Promise<any> => undefined,
-	saveHostMetric: async (..._args: any[]): Promise<any> => undefined,
-	getHostMetrics: async (..._args: any[]): Promise<any> => undefined,
-	getLatestHostMetrics: async (..._args: any[]): Promise<any> => undefined,
-	getAutoUpdateSettings: async (..._args: any[]): Promise<any> => undefined,
-	getAutoUpdateSetting: async (..._args: any[]): Promise<any> => undefined,
-	getAutoUpdateSettingById: async (..._args: any[]): Promise<any> => undefined,
-	updateAutoUpdateSettingById: async (..._args: any[]): Promise<any> => undefined,
-	getEnabledAutoUpdateSettings: async (..._args: any[]): Promise<any> => undefined,
-	getAllAutoUpdateSettings: async (..._args: any[]): Promise<any> => undefined,
-	upsertAutoUpdateSetting: async (..._args: any[]): Promise<any> => undefined,
-	updateAutoUpdateLastChecked: async (..._args: any[]): Promise<any> => undefined,
-	updateAutoUpdateLastUpdated: async (..._args: any[]): Promise<any> => undefined,
-	deleteAutoUpdateSetting: async (..._args: any[]): Promise<any> => undefined,
-	deleteAutoUpdateSchedule: async (..._args: any[]): Promise<any> => undefined,
-	renameAutoUpdateSchedule: async (..._args: any[]): Promise<any> => undefined,
-	NOTIFICATION_EVENT_TYPES: [] as any[],
-	NOTIFICATION_EVENT_GROUPS: [] as any[],
-	SYSTEM_NOTIFICATION_EVENTS: [] as any[],
-	ENVIRONMENT_NOTIFICATION_EVENTS: [] as any[],
-	getNotificationSettings: async (..._args: any[]): Promise<any> => undefined,
-	getNotificationSetting: async (..._args: any[]): Promise<any> => undefined,
-	getEnabledNotificationSettings: async (..._args: any[]): Promise<any> => undefined,
-	createNotificationSetting: async (..._args: any[]): Promise<any> => undefined,
-	updateNotificationSetting: async (..._args: any[]): Promise<any> => undefined,
-	deleteNotificationSetting: async (..._args: any[]): Promise<any> => undefined,
-	getEnvironmentNotifications: async (..._args: any[]): Promise<any> => undefined,
-	getEnvironmentNotification: async (..._args: any[]): Promise<any> => undefined,
-	createEnvironmentNotification: async (..._args: any[]): Promise<any> => undefined,
-	updateEnvironmentNotification: async (..._args: any[]): Promise<any> => undefined,
-	deleteEnvironmentNotification: async (..._args: any[]): Promise<any> => undefined,
-	getEnabledEnvironmentNotifications: async (..._args: any[]): Promise<any> => undefined,
-	getAuthSettings: async (..._args: any[]): Promise<any> => undefined,
-	updateAuthSettings: async (..._args: any[]): Promise<any> => undefined,
-	getUsers: async (..._args: any[]): Promise<any> => undefined,
-	getUser: async (..._args: any[]): Promise<any> => undefined,
-	getUserWithoutPassword: async (..._args: any[]): Promise<any> => undefined,
-	hasAdminUser: async (..._args: any[]): Promise<any> => undefined,
-	hasAnyUser: async (..._args: any[]): Promise<any> => undefined,
-	countAdminUsers: async (..._args: any[]): Promise<any> => undefined,
-	getUserByUsername: async (..._args: any[]): Promise<any> => undefined,
-	createUser: async (..._args: any[]): Promise<any> => undefined,
-	updateUser: async (..._args: any[]): Promise<any> => undefined,
-	deleteUser: async (..._args: any[]): Promise<any> => undefined,
-	createSession: async (..._args: any[]): Promise<any> => undefined,
-	getSession: async (..._args: any[]): Promise<any> => undefined,
-	deleteSession: async (..._args: any[]): Promise<any> => undefined,
-	deleteExpiredSessions: async (..._args: any[]): Promise<any> => undefined,
-	deleteUserSessions: async (..._args: any[]): Promise<any> => undefined,
-	getRoles: async (..._args: any[]): Promise<any> => undefined,
-	getRole: async (..._args: any[]): Promise<any> => undefined,
-	getRoleByName: async (..._args: any[]): Promise<any> => undefined,
-	createRole: async (..._args: any[]): Promise<any> => undefined,
-	updateRole: async (..._args: any[]): Promise<any> => undefined,
-	deleteRole: async (..._args: any[]): Promise<any> => undefined,
-	getUserRoles: async (..._args: any[]): Promise<any> => undefined,
-	assignUserRole: async (..._args: any[]): Promise<any> => undefined,
-	removeUserRole: async (..._args: any[]): Promise<any> => undefined,
-	userHasAdminRole: async (..._args: any[]): Promise<any> => undefined,
-	getUserAccessibleEnvironments: async (..._args: any[]): Promise<any> => undefined,
-	getUserRolesForEnvironment: async (..._args: any[]): Promise<any> => undefined,
-	userCanAccessEnvironment: async (..._args: any[]): Promise<any> => undefined,
-	getLdapConfigs: async (..._args: any[]): Promise<any> => undefined,
-	getLdapConfig: async (..._args: any[]): Promise<any> => undefined,
-	createLdapConfig: async (..._args: any[]): Promise<any> => undefined,
-	updateLdapConfig: async (..._args: any[]): Promise<any> => undefined,
-	deleteLdapConfig: async (..._args: any[]): Promise<any> => undefined,
-	getOidcConfigs: async (..._args: any[]): Promise<any> => undefined,
-	getOidcConfig: async (..._args: any[]): Promise<any> => undefined,
-	createOidcConfig: async (..._args: any[]): Promise<any> => undefined,
-	updateOidcConfig: async (..._args: any[]): Promise<any> => undefined,
-	deleteOidcConfig: async (..._args: any[]): Promise<any> => undefined,
-	getGitCredentials: async (..._args: any[]): Promise<any> => undefined,
-	getGitCredential: async (..._args: any[]): Promise<any> => undefined,
-	createGitCredential: async (..._args: any[]): Promise<any> => undefined,
-	updateGitCredential: async (..._args: any[]): Promise<any> => undefined,
-	deleteGitCredential: async (..._args: any[]): Promise<any> => undefined,
-	getGitRepositories: async (..._args: any[]): Promise<any> => undefined,
-	getGitRepository: async (..._args: any[]): Promise<any> => undefined,
-	getGitRepositoryByName: async (..._args: any[]): Promise<any> => undefined,
-	createGitRepository: async (..._args: any[]): Promise<any> => undefined,
-	updateGitRepository: async (..._args: any[]): Promise<any> => undefined,
-	getGitStacksByRepositoryId: async (..._args: any[]): Promise<any> => undefined,
-	deleteGitRepository: async (..._args: any[]): Promise<any> => undefined,
-	getGitStacks: async (..._args: any[]): Promise<any> => undefined,
-	getGitStacksForEnvironmentOnly: async (..._args: any[]): Promise<any> => undefined,
-	getGitStack: async (..._args: any[]): Promise<any> => undefined,
-	getGitStackByName: async (..._args: any[]): Promise<any> => undefined,
-	getGitStackByWebhookSecret: async (..._args: any[]): Promise<any> => undefined,
-	createGitStack: async (..._args: any[]): Promise<any> => undefined,
-	updateGitStack: async (..._args: any[]): Promise<any> => undefined,
-	deleteGitStack: async (..._args: any[]): Promise<any> => undefined,
-	renameGitStack: async (..._args: any[]): Promise<any> => undefined,
-	getEnabledAutoUpdateGitStacks: async (..._args: any[]): Promise<any> => undefined,
-	getAllAutoUpdateGitStacks: async (..._args: any[]): Promise<any> => undefined,
-	getStackSource: async (..._args: any[]): Promise<any> => undefined,
-	getStackSourceByComposePath: async (..._args: any[]): Promise<any> => undefined,
-	getStackSources: async (..._args: any[]): Promise<any> => undefined,
-	upsertStackSource: async (..._args: any[]): Promise<any> => undefined,
-	updateStackSource: async (..._args: any[]): Promise<any> => undefined,
-	setStackInjectedSecretKeys: async (..._args: any[]): Promise<any> => undefined,
-	getStackInjectedSecretKeys: async (..._args: any[]): Promise<any> => undefined,
-	deleteStackSource: async (..._args: any[]): Promise<any> => undefined,
-	updateStackSourceName: async (..._args: any[]): Promise<any> => undefined,
-	getContainerIconOverride: async (..._args: any[]): Promise<any> => undefined,
-	getContainerIconOverrides: async (..._args: any[]): Promise<any> => undefined,
-	setContainerIconOverride: async (..._args: any[]): Promise<any> => undefined,
-	deleteContainerIconOverride: async (..._args: any[]): Promise<any> => undefined,
-	saveVulnerabilityScan: async (..._args: any[]): Promise<any> => undefined,
-	getVulnerabilityScan: async (..._args: any[]): Promise<any> => undefined,
-	getLatestScanForImage: async (..._args: any[]): Promise<any> => undefined,
-	deleteScansForImageScanner: async (..._args: any[]): Promise<any> => undefined,
-	getScansForImage: async (..._args: any[]): Promise<any> => undefined,
-	getCombinedScanForImage: async (..._args: any[]): Promise<any> => undefined,
-	getAllLatestScans: async (..._args: any[]): Promise<any> => undefined,
-	getScanFreshness: async (..._args: any[]): Promise<any> => undefined,
-	deleteOldScans: async (..._args: any[]): Promise<any> => undefined,
-	logAuditEvent: async (..._args: any[]): Promise<any> => undefined,
-	getAuditLog: async (..._args: any[]): Promise<any> => undefined,
-	getAuditLogs: async (..._args: any[]): Promise<any> => undefined,
-	getAuditLogUsers: async (..._args: any[]): Promise<any> => undefined,
-	deleteOldAuditLogs: async (..._args: any[]): Promise<any> => undefined,
-	logContainerEvent: async (..._args: any[]): Promise<any> => undefined,
-	getContainerEvent: async (..._args: any[]): Promise<any> => undefined,
-	getContainerEvents: async (..._args: any[]): Promise<any> => undefined,
-	getContainerEventContainers: async (..._args: any[]): Promise<any> => undefined,
-	getContainerEventActions: async (..._args: any[]): Promise<any> => undefined,
-	deleteOldContainerEvents: async (..._args: any[]): Promise<any> => undefined,
-	runVolumeHelperCleanup: async (..._args: any[]): Promise<any> => undefined,
-	clearContainerEvents: async (..._args: any[]): Promise<any> => undefined,
-	getContainerEventStats: async (..._args: any[]): Promise<any> => undefined,
-	getUserPreference: async (..._args: any[]): Promise<any> => undefined,
-	setUserPreference: async (..._args: any[]): Promise<any> => undefined,
-	deleteUserPreference: async (..._args: any[]): Promise<any> => undefined,
-	getDashboardPreferences: async (..._args: any[]): Promise<any> => undefined,
-	saveDashboardPreferences: async (..._args: any[]): Promise<any> => undefined,
-	createScheduleExecution: async (..._args: any[]): Promise<any> => undefined,
-	updateScheduleExecution: async (..._args: any[]): Promise<any> => undefined,
-	failStaleRunningExecutions: async (..._args: any[]): Promise<any> => undefined,
-	appendScheduleExecutionLog: async (..._args: any[]): Promise<any> => undefined,
-	getScheduleExecution: async (..._args: any[]): Promise<any> => undefined,
-	deleteScheduleExecution: async (..._args: any[]): Promise<any> => undefined,
-	getScheduleExecutions: async (..._args: any[]): Promise<any> => undefined,
-	getScheduleStats: async (..._args: any[]): Promise<any> => undefined,
-	getBackupConfigMetrics: async (..._args: any[]): Promise<any> => undefined,
-	getLastExecutionForSchedule: async (..._args: any[]): Promise<any> => undefined,
-	getRecentExecutionsForSchedule: async (..._args: any[]): Promise<any> => undefined,
-	cleanupOldExecutions: async (..._args: any[]): Promise<any> => undefined,
-	getScheduleRetentionDays: async (..._args: any[]): Promise<any> => undefined,
-	setScheduleRetentionDays: async (..._args: any[]): Promise<any> => undefined,
-	getEventRetentionDays: async (..._args: any[]): Promise<any> => undefined,
-	setEventRetentionDays: async (..._args: any[]): Promise<any> => undefined,
-	getScheduleCleanupCron: async (..._args: any[]): Promise<any> => undefined,
-	setScheduleCleanupCron: async (..._args: any[]): Promise<any> => undefined,
-	getEventCleanupCron: async (..._args: any[]): Promise<any> => undefined,
-	setEventCleanupCron: async (..._args: any[]): Promise<any> => undefined,
-	getScheduleCleanupEnabled: async (..._args: any[]): Promise<any> => undefined,
-	setScheduleCleanupEnabled: async (..._args: any[]): Promise<any> => undefined,
-	getEventCleanupEnabled: async (..._args: any[]): Promise<any> => undefined,
-	setEventCleanupEnabled: async (..._args: any[]): Promise<any> => undefined,
-	getScannerCleanupCron: async (..._args: any[]): Promise<any> => undefined,
-	setScannerCleanupCron: async (..._args: any[]): Promise<any> => undefined,
-	getScannerCleanupEnabled: async (..._args: any[]): Promise<any> => undefined,
-	setScannerCleanupEnabled: async (..._args: any[]): Promise<any> => undefined,
-	getExternalStackPaths: async (..._args: any[]): Promise<any> => undefined,
-	setExternalStackPaths: async (..._args: any[]): Promise<any> => undefined,
-	addExternalStackPath: async (..._args: any[]): Promise<any> => undefined,
-	getPrimaryStackLocation: async (..._args: any[]): Promise<any> => undefined,
-	setPrimaryStackLocation: async (..._args: any[]): Promise<any> => undefined,
-	getGlobalSemverConfig: async (..._args: any[]): Promise<any> => undefined,
-	setGlobalSemverConfig: async (..._args: any[]): Promise<any> => undefined,
-	getEnvUpdateCheckSettings: async (..._args: any[]): Promise<any> => undefined,
-	setEnvUpdateCheckSettings: async (..._args: any[]): Promise<any> => undefined,
-	deleteEnvUpdateCheckSettings: async (..._args: any[]): Promise<any> => undefined,
-	getAllEnvUpdateCheckSettings: async (..._args: any[]): Promise<any> => undefined,
-	getImagePruneSettings: async (..._args: any[]): Promise<any> => undefined,
-	setImagePruneSettings: async (..._args: any[]): Promise<any> => undefined,
-	deleteImagePruneSettings: async (..._args: any[]): Promise<any> => undefined,
-	getAllImagePruneSettings: async (..._args: any[]): Promise<any> => undefined,
-	getEnvironmentTimezone: async (..._args: any[]): Promise<any> => undefined,
-	setEnvironmentTimezone: async (..._args: any[]): Promise<any> => undefined,
-	getDefaultTimezone: async (..._args: any[]): Promise<any> => undefined,
-	setDefaultTimezone: async (..._args: any[]): Promise<any> => undefined,
-	getEventCollectionMode: async (..._args: any[]): Promise<any> => undefined,
-	setEventCollectionMode: async (..._args: any[]): Promise<any> => undefined,
-	getEventPollInterval: async (..._args: any[]): Promise<any> => undefined,
-	setEventPollInterval: async (..._args: any[]): Promise<any> => undefined,
-	getMetricsCollectionInterval: async (..._args: any[]): Promise<any> => undefined,
-	setMetricsCollectionInterval: async (..._args: any[]): Promise<any> => undefined,
-	getStackEnvVars: async (..._args: any[]): Promise<any> => undefined,
-	getStackEnvVarsAsRecord: async (..._args: any[]): Promise<any> => undefined,
-	getSecretEnvVarsAsRecord: async (..._args: any[]): Promise<any> => undefined,
-	getStackSecretCiphertexts: async (..._args: any[]): Promise<any> => undefined,
-	getNonSecretEnvVarsAsRecord: async (..._args: any[]): Promise<any> => undefined,
-	setStackEnvVars: async (..._args: any[]): Promise<any> => undefined,
-	getSecretKeyNames: async (..._args: any[]): Promise<any> => undefined,
-	getSecretKeysToMask: async (..._args: any[]): Promise<any> => undefined,
-	getStackEnvVarsCount: async (..._args: any[]): Promise<any> => undefined,
-	deleteStackEnvVars: async (..._args: any[]): Promise<any> => undefined,
-	updateStackEnvVarsName: async (..._args: any[]): Promise<any> => undefined,
-	getAllStacksEnvVarsCounts: async (..._args: any[]): Promise<any> => undefined,
-	getPendingContainerUpdates: async (..._args: any[]): Promise<any> => undefined,
-	clearPendingContainerUpdates: async (..._args: any[]): Promise<any> => undefined,
-	addPendingContainerUpdate: async (..._args: any[]): Promise<any> => undefined,
-	removePendingContainerUpdate: async (..._args: any[]): Promise<any> => undefined,
-	getBackupDestinations: async (..._args: any[]): Promise<any> => undefined,
-	getBackupDestination: async (..._args: any[]): Promise<any> => undefined,
-	createBackupDestination: async (..._args: any[]): Promise<any> => undefined,
-	updateBackupDestination: async (..._args: any[]): Promise<any> => undefined,
-	deleteBackupDestination: async (..._args: any[]): Promise<any> => undefined,
-	updateBackupDestinationTestStatus: async (..._args: any[]): Promise<any> => undefined,
-	decryptBackupDestination: async (..._args: any[]): Promise<any> => undefined,
-	getBackupConfigs: async (..._args: any[]): Promise<any> => undefined,
-	getBackupConfig: async (..._args: any[]): Promise<any> => undefined,
-	getEnabledBackupConfigs: async (..._args: any[]): Promise<any> => undefined,
-	createBackupConfig: async (..._args: any[]): Promise<any> => undefined,
-	updateBackupConfig: async (..._args: any[]): Promise<any> => undefined,
-	deleteBackupConfig: async (..._args: any[]): Promise<any> => undefined,
-	db: {} as any,
-	isPostgres: false,
-	isSqlite: true,
-	parseEnvInterpolation: (s: string) => s,
-}));
-
-const { collectProcess } = await import('../src/lib/server/stacks');
+import { collectProcess } from '../src/lib/server/process-output-core';
 
 describe('collectProcess mit Zeilen-Rueckruf', () => {
 	test('meldet jede Zeile einzeln und in der Reihenfolge', async () => {
